// models/DelegateSubmissionBody.ts
/**
 * vegawallet-service-client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DelegateSubmissionBodyDelegateSubmission } from './DelegateSubmissionBodyDelegateSubmission';
import { HttpFile } from '../http/http';

export class DelegateSubmissionBody {
    'pubKey': string;
    'propagate': boolean;
    'delegateSubmission': DelegateSubmissionBodyDelegateSubmission;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "pubKey",
            "baseName": "pubKey",
            "type": "string",
            "format": ""
        },
        {
            "name": "propagate",
            "baseName": "propagate",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "delegateSubmission",
            "baseName": "delegateSubmission",
            "type": "DelegateSubmissionBodyDelegateSubmission",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DelegateSubmissionBody.attributeTypeMap;
    }

    public constructor() {
    }
}