// models/GenericError.ts
/**
 * vegawallet-service-client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GenericErrorErrors } from './GenericErrorErrors';
import { HttpFile } from '../http/http';

export class GenericError {
    'errors': GenericErrorErrors;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "errors",
            "baseName": "errors",
            "type": "GenericErrorErrors",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GenericError.attributeTypeMap;
    }

    public constructor() {
    }
}