// models/KeysGet200Response.ts
/**
 * vegawallet-service-client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { VegaKey } from './VegaKey';
import { HttpFile } from '../http/http';

export class KeysGet200Response {
    'keys': Array<VegaKey>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "keys",
            "baseName": "keys",
            "type": "Array<VegaKey>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return KeysGet200Response.attributeTypeMap;
    }

    public constructor() {
    }
}