// models/NetworkConfigApi.ts
/**
 * vegawallet-service-client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { NetworkConfigApiGrpc } from './NetworkConfigApiGrpc';
import { NetworkConfigApiRest } from './NetworkConfigApiRest';
import { HttpFile } from '../http/http';

export class NetworkConfigApi {
    'grpc'?: NetworkConfigApiGrpc;
    'rest'?: NetworkConfigApiRest;
    'graphQl'?: NetworkConfigApiRest;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "grpc",
            "baseName": "grpc",
            "type": "NetworkConfigApiGrpc",
            "format": ""
        },
        {
            "name": "rest",
            "baseName": "rest",
            "type": "NetworkConfigApiRest",
            "format": ""
        },
        {
            "name": "graphQl",
            "baseName": "graphQl",
            "type": "NetworkConfigApiRest",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return NetworkConfigApi.attributeTypeMap;
    }

    public constructor() {
    }
}