// models/NetworkConfigApiGrpc.ts
/**
 * vegawallet-service-client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class NetworkConfigApiGrpc {
    'hosts'?: Array<string>;
    'retries'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "hosts",
            "baseName": "hosts",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "retries",
            "baseName": "retries",
            "type": "number",
            "format": "int32"
        }    ];

    static getAttributeTypeMap() {
        return NetworkConfigApiGrpc.attributeTypeMap;
    }

    public constructor() {
    }
}